+++ java_design_pattern/src/at/ac/htlinn/design_pattern/strategy/payment/PaymentStrategy.java
@@ -4,5 +4,5 @@ package at.ac.htlinn.design_pattern.strategy.payment;
  * based on: https://www.journaldev.com/1754/strategy-design-pattern-in-java-example-tutorial
  */
 public interface PaymentStrategy {
-	public void pay(int amount);
+	public void pay(double amount);
 }
