+++ java_design_pattern/src/at/ac/htlinn/design_pattern/strategy/payment/ShoppingCart.java
@@ -19,8 +19,8 @@ public class ShoppingCart {
 		this.items.remove(item);
 	}
 
-	public int calculateTotal() {
-		int sum = 0;
+	public double calculateTotal() {
+		double sum = 0;
 		for (Article item : items) {
 			sum += item.getPrice();
 		}
@@ -28,7 +28,7 @@ public class ShoppingCart {
 	}
 
 	public void pay(PaymentStrategy paymentMethod) {
-		int amount = calculateTotal();
+		double amount = calculateTotal();
 		paymentMethod.pay(amount);
 	}
